+++ inc/class.yoast-compatibility.php
@@ -12,7 +12,8 @@ class YOASTCompatibility extends Singleton {
   private $prefix;
 
   public function __construct(SEO $seo) {
-    add_action( 'plugins_loaded', [$this, 'disable_yoast_seo_frontend'] );
+    add_action( 'plugins_loaded', [$this, 'disable_yoast_frontend'] );
+    add_action( 'plugins_loaded', [$this, 'disable_taxonomy_metaboxes'] );
     add_action( 'add_meta_boxes', [$this, 'meta_boxes'], 11 );
   }
 
@@ -21,12 +22,18 @@ class YOASTCompatibility extends Singleton {
    *
    * @return void
    */
-  public function disable_yoast_seo_frontend() {
+  public function disable_yoast_frontend() {
     if( is_admin() || !defined('WPSEO_VERSION') ) return;
     $loader = \YoastSEO()->classes->get( \Yoast\WP\SEO\Loader::class );
     \remove_action( 'init', [ $loader, 'load_integrations' ] );
     \remove_action( 'rest_api_init', [ $loader, 'load_routes' ] );
   }
+
+
+  public function disable_taxonomy_metaboxes() {
+    
+  }
+  
 
   /**
    * Remove the YOAST Meta Box everywhere
@@ -36,6 +43,8 @@ class YOASTCompatibility extends Singleton {
   public function meta_boxes() {
     remove_meta_box( 'wpseo_meta', null, 'normal' );
     remove_meta_box( 'wpseo_meta', null, 'side' );
+
+    
   }
 
 }