+++ changelog.md
@@ -1,6 +1,7 @@
 #### 1.0.2 (2021-02-01)
 
-- Compatibility with YOAST SEO (#05dec04)
+- Support for terms, deeper integration with qTranslate (#8e1f474)
+- Compatibility with YOAST SEO (#c9c5f24)
 
 #### 1.0.1 (2020-12-08)